pyproject.toml
[project]
name = "neo-rag-research-agent"
version = "0.1.0"
description = "A RAG research agent using LangGraph"
authors = [
    { name = "Neo", email = "espen@neo.no" }
]
dependencies = [
    "langgraph>=0.0.19",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-pinecone>=0.0.2",
    "pinecone-client>=3.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
extend-select = ["I"]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.entry-points."langgraph.templates"]
"neo_rag_agent" = "src"

[project.entry-points."langgraph.graphs"]
"neo_rag_agent" = "src"